# Allow serde and serde_derive to compile in parallel