--- a/Cargo.toml
+++ b/Cargo.toml
@@ -30,7 +30,8 @@ log_secure_cookie_values = []
 idna = "0.3.0"
 log = "0.4.17"
 publicsuffix = { version = "2.2.3", optional = true }
-serde = { version = "1.0.147", features = [ "derive" ] }
+serde = "1.0.147"
+serde_derive = "1.0.147"
 serde_json = "1.0.87"
 time = "0.3.16"
 url = "2.3.1"