--- a/src/cookie_path.rs
+++ b/src/cookie_path.rs
@@ -1,4 +1,4 @@
-use serde::{Deserialize, Serialize};
+use serde_derive::{Deserialize, Serialize};
 use std::cmp::max;
 use std::ops::Deref;
 use url::Url;