Visualized in `cargo build --timings`:

| Before: | After: |
|---|---|
| <img src="https://github.com/pfernie/cookie_store/assets/1940490/4d1893b4-2fe5-446d-b88c-1f3fd751b482" width="350"> | <img src="https://github.com/pfernie/cookie_store/assets/1940490/47409b0d-3050-495c-8cf1-f486f8f1ab0a" width="350"> |

Previously `serde` and `serde_json` could only begin compiling after `serde_derive` was finished compiling. After, they can compile in parallel.

This reduces build time by 2 seconds (30%). This crate is in the critical path of the widely used `reqwest` crate, so I think this is worth doing.