@pfernie 2023-06-17
Thanks for the PR! This is merged, and also backported to the 0.16 (current version included from `reqwest`) branch. Released as `0.16.2` and `0.19.1`.